+++ docs/conf.py
@@ -27,7 +27,7 @@ copyright = "2022, Quantmetry"
 author = "Quantmetry"
 
 # The full version, including alpha/beta/rc tags
-version = "0.1.2"
+version = "0.1.6"
 release = version
 
 # -- General configuration ---------------------------------------------------