+++ qolmat/imputations/em_sampler.py
@@ -71,7 +71,7 @@ def _conjugate_gradient(A: NDArray, X: NDArray, mask: NDArray) -> NDArray:
     return X_final
 
 
-def min_diff_Linf(list_params: List[NDArray], n_steps: int, order: int = 1) -> float:
+def max_diff_Linf(list_params: List[NDArray], n_steps: int, order: int = 1) -> float:
     """Computes the maximal L infinity norm between the `n_steps` last elements spaced by order.
     Used to compute the stop criterion.
 
@@ -466,14 +466,14 @@ class EM(BaseEstimator, TransformerMixin):
         IllConditioned
             Data matrix is ill-conditioned due to colinear columns.
         """
-        n_rows, n_cols = X.shape
+        n_samples, n_cols = X.shape
         # if n_rows == 1 the function np.cov returns a float
-        if n_rows == 1:
-            min_sv = 0
-        else:
-            cov = np.cov(X, bias=True, rowvar=False).reshape(n_cols, -1)
-            _, sv, _ = spl.svd(cov)
-            min_sv = min(np.sqrt(sv))
+        if n_samples == 1:
+            raise ValueError("EM cannot be fitted when n_samples = 1!")
+
+        cov = np.cov(X, bias=True, rowvar=False).reshape(n_cols, -1)
+        _, sv, _ = spl.svd(cov)
+        min_sv = min(np.sqrt(sv))
         if min_sv < self.min_std:
             warnings.warn(
                 f"The covariance matrix is ill-conditioned, indicating high-colinearity: the "
@@ -481,7 +481,6 @@ class EM(BaseEstimator, TransformerMixin):
                 f"min_std ({min_sv} < {self.min_std}). Consider removing columns of decreasing "
                 f"the threshold."
             )
-            # raise IllConditioned(min_sv, self.min_std)
 
 
 class MultiNormalEM(EM):
@@ -683,8 +682,7 @@ class MultiNormalEM(EM):
         X : NDArray
             Data matrix with missingness
         """
-        self.means = np.nanmean(X, axis=0)
-        self.cov = utils.nancov(X)
+        self.means, self.cov = utils.nan_mean_cov(X)
         self.cov_inv = np.linalg.pinv(self.cov)
 
     def set_parameters(self, means: NDArray, cov: NDArray):
@@ -764,8 +762,8 @@ class MultiNormalEM(EM):
         if n_iter < 3:
             return False
 
-        min_diff_means1 = min_diff_Linf(list_covs, n_steps=1)
-        min_diff_covs1 = min_diff_Linf(list_means, n_steps=1)
+        min_diff_means1 = max_diff_Linf(list_means, n_steps=1)
+        min_diff_covs1 = max_diff_Linf(list_covs, n_steps=1)
         min_diff_reached = min_diff_means1 < self.tolerance and min_diff_covs1 < self.tolerance
 
         if min_diff_reached:
@@ -774,16 +772,16 @@ class MultiNormalEM(EM):
         if n_iter < 7:
             return False
 
-        min_diff_means5 = min_diff_Linf(list_covs, n_steps=5)
-        min_diff_covs5 = min_diff_Linf(list_means, n_steps=5)
+        min_diff_means5 = max_diff_Linf(list_means, n_steps=5)
+        min_diff_covs5 = max_diff_Linf(list_covs, n_steps=5)
 
         min_diff_stable = (
             min_diff_means5 < self.stagnation_threshold
             and min_diff_covs5 < self.stagnation_threshold
         )
 
-        min_diff_loglik5_ord1 = min_diff_Linf(list_logliks, n_steps=5)
-        min_diff_loglik5_ord2 = min_diff_Linf(list_logliks, n_steps=5, order=2)
+        min_diff_loglik5_ord1 = max_diff_Linf(list_logliks, n_steps=5)
+        min_diff_loglik5_ord2 = max_diff_Linf(list_logliks, n_steps=5, order=2)
         max_loglik = (min_diff_loglik5_ord1 < self.stagnation_loglik) or (
             min_diff_loglik5_ord2 < self.stagnation_loglik
         )
@@ -1107,8 +1105,8 @@ class VARpEM(EM):
         if n_iter < 3:
             return False
 
-        min_diff_B1 = min_diff_Linf(list_B, n_steps=1)
-        min_diff_S1 = min_diff_Linf(list_S, n_steps=1)
+        min_diff_B1 = max_diff_Linf(list_B, n_steps=1)
+        min_diff_S1 = max_diff_Linf(list_S, n_steps=1)
         min_diff_reached = min_diff_B1 < self.tolerance and min_diff_S1 < self.tolerance
 
         if min_diff_reached:
@@ -1117,14 +1115,14 @@ class VARpEM(EM):
         if n_iter < 7:
             return False
 
-        min_diff_B5 = min_diff_Linf(list_B, n_steps=5)
-        min_diff_S5 = min_diff_Linf(list_S, n_steps=5)
+        min_diff_B5 = max_diff_Linf(list_B, n_steps=5)
+        min_diff_S5 = max_diff_Linf(list_S, n_steps=5)
         min_diff_stable = (
             min_diff_B5 < self.stagnation_threshold and min_diff_S5 < self.stagnation_threshold
         )
 
-        max_loglik5_ord1 = min_diff_Linf(list_logliks, n_steps=5, order=1)
-        max_loglik5_ord2 = min_diff_Linf(list_logliks, n_steps=5, order=2)
+        max_loglik5_ord1 = max_diff_Linf(list_logliks, n_steps=5, order=1)
+        max_loglik5_ord2 = max_diff_Linf(list_logliks, n_steps=5, order=2)
         max_loglik = (max_loglik5_ord1 < self.stagnation_loglik) or (
             max_loglik5_ord2 < self.stagnation_loglik
         )