+++ qolmat/utils/data.py
@@ -3,7 +3,7 @@ import sys
 import zipfile
 from datetime import datetime
 from math import pi
-from typing import List
+from typing import List, Tuple, Union
 from urllib import request
 
 import numpy as np
@@ -15,26 +15,45 @@ CURRENT_DIR = os.path.abspath(os.path.dirname(__file__))
 ROOT_DIR = os.path.join(CURRENT_DIR, "..")
 
 
-def read_csv_local(data_file_name: str) -> pd.DataFrame:
+def read_csv_local(data_file_name: str, **kwargs) -> pd.DataFrame:
     """Load csv files
 
     Parameters
     ----------
     data_file_name : str
         Filename. Has to be "beijing" or "conductors"
+    kwargs : dict
 
     Returns
     -------
     df : pd.DataFrame
         dataframe
     """
-    df = pd.read_csv(os.path.join(ROOT_DIR, "data", f"{data_file_name}.csv"))
+    df = pd.read_csv(os.path.join(ROOT_DIR, "data", f"{data_file_name}.csv"), **kwargs)
     return df
 
 
 def download_data_from_zip(
     zipname: str, urllink: str, datapath: str = "data/"
 ) -> List[pd.DataFrame]:
+    """
+    Downloads and extracts ZIP files from a URL, then loads DataFrames from CSV files.
+
+    Parameters
+    ----------
+    zipname : str
+        Name of the ZIP file to download, without the '.zip' extension.
+    urllink : str
+        Base URL where the ZIP file is hosted.
+    datapath : str, optional
+        Path to the directory where the ZIP will be downloaded and extracted.
+        Defaults to 'data/'.
+
+    Returns
+    -------
+    List[pd.DataFrame]
+        A list of DataFrames loaded from the CSV files within the extracted directory.
+    """
     path_zip = os.path.join(datapath, zipname)
     path_zip_ext = path_zip + ".zip"
     url = os.path.join(urllink, zipname) + ".zip"
@@ -49,6 +68,23 @@ def download_data_from_zip(
 
 
 def get_dataframes_in_folder(path: str, extension: str) -> List[pd.DataFrame]:
+    """
+    Loads all dataframes from files with a specified extension within a directory, including
+    subdirectories. Special handling for '.tsf' files which are converted and immediately returned.
+
+    Parameters
+    ----------
+    path : str
+        Path to the directory to search for files.
+    extension : str
+        File extension to filter files by, e.g., '.csv'.
+
+    Returns
+    -------
+    List[pd.DataFrame]
+        A list of pandas DataFrames loaded from the files matching the extension.
+        If a '.tsf' file is found, its converted DataFrame is returned immediately.
+    """
     list_df = []
     for folder, _, files in os.walk(path):
         for file in files:
@@ -60,7 +96,37 @@ def get_dataframes_in_folder(path: str, extension: str) -> List[pd.DataFrame]:
     return list_df
 
 
-def generate_artificial_ts(n_samples, periods, amp_anomalies, ratio_anomalies, amp_noise):
+def generate_artificial_ts(
+    n_samples: int,
+    periods: List[int],
+    amp_anomalies: float,
+    ratio_anomalies: float,
+    amp_noise: float,
+) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Generates time series data, anomalies, and noise based on given parameters.
+
+    Parameters
+    ----------
+    n_samples : int
+        Number of samples in the time series.
+    periods : List[int]
+        List of periods that are added to the time series.
+    amp_anomalies : float
+        Amplitude multiplier for anomalies.
+    ratio_anomalies : float
+        Ratio of total samples that will be anomalies.
+    amp_noise : float
+        Standard deviation of Gaussian noise.
+
+    Returns
+    -------
+    Tuple[np.ndarray, np.ndarray, np.ndarray]
+        Time series data with sine waves (X).
+        Anomaly data with specified amplitudes at random positions (A).
+        Gaussian noise added to the time series (E).
+    """
+
     mesh = np.arange(n_samples)
     X = np.ones(n_samples)
     for p in periods:
@@ -82,7 +148,8 @@ def get_data(
     datapath: str = "data/",
     n_groups_max: int = sys.maxsize,
 ) -> pd.DataFrame:
-    """Download or generate data
+    """
+    Download or generate data
 
     Parameters
     ----------
@@ -101,21 +168,19 @@ def get_data(
     if name_data == "Beijing":
         df = read_csv_local("beijing")
         df["date"] = pd.to_datetime(df["date"])
-
-        # df["date"] = pd.to_datetime(
-        #     {
-        #         "year": df["year"],
-        #         "month": df["month"],
-        #         "day": df["day"],
-        #         "hour": df["hour"],
-        #     }
-        # )
         df = df.drop(columns=["year", "month", "day", "hour", "wd"])
-        # df = df.set_index(["station", "date"])
         df = df.groupby(["station", "date"]).mean()
         return df
-    if name_data == "Superconductor":
+    elif name_data == "Superconductor":
         df = read_csv_local("conductors")
+        return df
+    elif name_data == "Titanic":
+        path = "https://gist.githubusercontent.com/fyyying/4aa5b471860321d7b47fd881898162b7/raw/"
+        "6907bb3a38bfbb6fccf3a8b1edfb90e39714d14f/titanic_dataset.csv"
+        df = pd.read_csv(path)
+        df = df[["Survived", "Sex", "Age", "SibSp", "Parch", "Fare", "Embarked"]]
+        df["Age"] = pd.to_numeric(df["Age"], errors="coerce")
+        df["Fare"] = pd.to_numeric(df["Fare"], errors="coerce")
         return df
     elif name_data == "Artificial":
         city = "Wonderland"
@@ -254,22 +319,16 @@ def add_holes(df: pd.DataFrame, ratio_masked: float, mean_size: int) -> pd.DataF
 
     ratio_masked : float
         Targeted global proportion of nans added in the returned dataset
-
-    groups: list of strings
-        List of the column names used as groups
-
     Returns
     -------
     pd.DataFrame
         dataframe with missing values
     """
-    try:
-        groups = df.index.names.difference(["datetime", "date", "index"])
+    groups = df.index.names.difference(["datetime", "date", "index", None])
+    if groups != []:
         generator = missing_patterns.GeometricHoleGenerator(
             1, ratio_masked=ratio_masked, subset=df.columns, groups=groups
         )
-    except ValueError:
-        print("No group")
     else:
         generator = missing_patterns.GeometricHoleGenerator(
             1, ratio_masked=ratio_masked, subset=df.columns
@@ -334,7 +393,7 @@ def add_station_features(df: pd.DataFrame) -> pd.DataFrame:
     return df
 
 
-def add_datetime_features(df: pd.DataFrame) -> pd.DataFrame:
+def add_datetime_features(df: pd.DataFrame, col_time: str = "datetime") -> pd.DataFrame:
     """
     Create a seasonal feature in the dataset with a cosine function
 
@@ -342,6 +401,8 @@ def add_datetime_features(df: pd.DataFrame) -> pd.DataFrame:
     ----------
     df : pd.DataFrame
         dataframe no missing values
+    col_time: string
+        Column of the index containing the time index
 
     Returns
     -------
@@ -349,12 +410,13 @@ def add_datetime_features(df: pd.DataFrame) -> pd.DataFrame:
         dataframe with missing values
     """
     df = df.copy()
-    time = df.index.get_level_values("datetime").to_series()
+    time = df.index.get_level_values(col_time).to_series()
     days_in_year = time.dt.year.apply(
         lambda x: 366 if ((x % 4 == 0) and (x % 100 != 0)) or (x % 400 == 0) else 365
     )
-    time_cos = np.cos(2 * np.pi * time.dt.dayofyear / days_in_year)
-    df["time_cos"] = np.array(time_cos)
+    ratio = time.dt.dayofyear.values / days_in_year.values
+    df["time_cos"] = np.cos(2 * np.pi * ratio)
+    df["time_sin"] = np.sin(2 * np.pi * ratio)
     return df
 
 
@@ -367,42 +429,27 @@ def convert_tsf_to_dataframe(
     col_types = []
     all_data = {}
     line_count = 0
-    # frequency = None
-    # forecast_horizon = None
-    # contain_missing_values = None
-    # contain_equal_length = None
     found_data_tag = False
     found_data_section = False
     started_reading_data_section = False
 
     with open(full_file_path_and_name, "r", encoding="cp1252") as file:
         for line in file:
-            # Strip white space from start/end of line
             line = line.strip()
 
             if line:
-                if line.startswith("@"):  # Read meta-data
+                if line.startswith("@"):
                     if not line.startswith("@data"):
                         line_content = line.split(" ")
                         if line.startswith("@attribute"):
-                            if len(line_content) != 3:  # Attributes have both name and type
+                            if len(line_content) != 3:
                                 raise Exception("Invalid meta-data specification.")
 
                             col_names.append(line_content[1])
                             col_types.append(line_content[2])
                         else:
-                            if len(line_content) != 2:  # Other meta-data have only values
+                            if len(line_content) != 2:
                                 raise Exception("Invalid meta-data specification.")
-
-                            # if line.startswith("@frequency"):
-                            #     frequency = line_content[1]
-                            # elif line.startswith("@horizon"):
-                            #     forecast_horizon = int(line_content[1])
-                            # elif line.startswith("@missing"):
-                            #     contain_missing_values = bool(strtobool(line_content[1]))
-                            # elif line.startswith("@equallength"):
-                            #     contain_equal_length = bool(strtobool(line_content[1]))
-
                     else:
                         if len(col_names) == 0:
                             raise Exception("Attribute section must come before data.")