+++ HISTORY.rst
@@ -2,11 +2,27 @@
 History
 =======
 
-??
+0.1.7 (2024-06-13)
 ------------------
 * Little's test implemented in a new hole_characterization module
 * Documentation now includes an analysis section with a tutorial
 * Hole generators now provide reproducible outputs
+
+0.1.5 (2024-04-17)
+------------------
+
+* CICD now relies on Node.js 20
+* New tests for comparator.py and data.py
+
+0.1.4 (2024-04-15)
+------------------
+
+* ImputerMean, ImputerMedian and ImputerMode have been merged into ImputerSimple
+* File preprocessing.py added with classes new MixteHGBM, BinTransformer, OneHotEncoderProjector and WrapperTransformer providing tools to manage mixed types data
+* Tutorial plot_tuto_categorical showcasing mixed type imputation
+* Titanic dataset added
+* accuracy metric implemented
+* metrics.py rationalized, and split with algebra.py
 
 0.1.3 (2024-03-07)
 ------------------